+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changed
 
 - Bump nfx-stringutils from 0.4.0 to 0.5.0
+- Remove unused variables from `std::from_chars` and `std::to_chars` calls
 
 ### Deprecated